removed performance mode, fixed bugs in FilePackage.move() and a lot of code quality improvements

+++ jsqp_core/packages/file_package.py
@@ -1,13 +1,17 @@
 from __future__ import annotations
+from typing import TYPE_CHECKING
 
 import os
 import shutil
-from io import FileIO
+import hashlib
 from enum import Enum
 from pathlib import Path
 from zipfile import ZipFile
 from abc import abstractmethod, ABC
 from devgoldyutils import LoggerAdapter
+
+if TYPE_CHECKING:
+    from io import BufferedReader
 
 from ..paths import Paths
 from .package import Package
@@ -76,12 +80,10 @@ class FilePackage(Package, ABC):
         self.logger.debug(f"Updated file package path object to '{str(value)}'!")
 
     @property
-    def file(self) -> FileIO | None:
+    def file(self) -> BufferedReader | None:
         """Returns the file object of this package. Returns none if this package isn't a file."""
         if self.__path.is_file():
-            file = open(str(self.__path), "r")
-            file.close()
-            return file
+            return open(str(self.__path), "rb")
 
         return None
 
@@ -100,24 +102,38 @@ class FilePackage(Package, ABC):
         elif self.path.is_symlink():
             return FileTypes.SYMBOLIC_LINK
 
+    @property
+    def hash(self) -> str | None:
+        """Returns the hash of the package's file. Returns None if self.file is none."""
+        file = self.file
+
+        if file is not None:
+            md5 = hashlib.md5()
+
+            with file as f:
+                md5.update(f.read())
+
+            return md5.hexdigest()
+
+        return None
+
     def add(self, overwrite: bool = False):
-        """Adds the package to the jsqp repository."""
-        self.logger.info(f"Adding '{self.display_name}' to jsqp repository...")
+        """Adds the package to your jsqp storage."""
+        self.logger.info(f"Adding '{self.display_name}' to jsqp 📦 storage...")
+
         # Zip it if it's not already a zip.
         if self.type == FileTypes.FOLDER:
-            self.zip(self.name + ".zip", config.performance_mode)
+            self.zip(self.name + ".zip")
 
         # Move the package to JSQPCore install location directory.
-        try:
-            self.move(self.package_install_path, overwrite, config.no_copy)
-        except FileNotFoundError:
-            # Repair and try again if file is not found.
-            # -------------------------------------------
-            Paths().repair_app_data_dir([
-                self.package_install_path
-            ])
 
-            self.move(self.package_install_path, overwrite, config.no_copy)
+        if not os.path.exists(self.package_install_path):
+            Paths().repair_app_data_dir([self.package_install_path])
+
+        try:
+            self.move(
+                self.package_install_path, overwrite, config.no_copy
+            )
         except FileExistsError:
             raise errors.PackageAlreadyExist(self)
 
@@ -141,19 +157,21 @@ class FilePackage(Package, ABC):
             self.logger.warning("Didn't create symbolic link as it already exists.")
 
     def move(self, path: str, overwrite: bool = False, no_copy: bool = True) -> None:
-        """Allows you to move this file package to another location. Raises ``FileNotFoundError`` if path or file does not exist."""
+        """
+        Allows you to move this file package to another location. 
+        Raises ``FileNotFoundError`` if path does not exist and ``FileExistsError`` if a file already exists there.
+        """
         new_path = f"{path}/{self.path.name}"
         self.logger.info(f"Moving '{self.display_name}' to '{os.path.split(new_path)[0]:.60}'...")
 
-        if overwrite or config.overwrite:
-            self.logger.info("Overwriting as it already exists...")
-            if os.path.isdir(new_path):
-                shutil.rmtree(new_path, True)
-            else:
-                os.remove(new_path)
-
         if os.path.exists(new_path):
-            raise FileExistsError("The file already exists.")
+
+            if overwrite or config.overwrite:
+                self.logger.info("Overwriting as it already exists...")
+                shutil.rmtree(new_path, True) if os.path.isdir(new_path) else os.remove(new_path)
+
+            else:
+                raise FileExistsError("The file already exists.")
 
         shutil.copy2(self.path.absolute(), new_path)
 
@@ -166,7 +184,7 @@ class FilePackage(Package, ABC):
 
         self.logger.info(f"Moved to '{new_path:.60}'!")
 
-    def zip(self, zip_name: str = None, performance_mode: bool = False) -> bool:
+    def zip(self, zip_name: str = None) -> bool:
         """Turn file package into a zip if it is a folder."""
         if zip_name is None:
             zip_name = self.name + ".zip"
@@ -176,22 +194,13 @@ class FilePackage(Package, ABC):
             path_to_zip = Path(str(self.path.parent) + f"/{zip_name}")
 
             self.logger.info(f"Zipping {self.display_name} to '{str(path_to_zip):.60}...'")
-            
-            # Zipping each file in directory.
-            # --------------------------------
-            with ZipFile(path_to_zip.absolute(), mode = "w") as archive:
 
-                if core_logger.level == 10 and performance_mode is False: # DEBUG MODE!
-                    for file_path in self.path.rglob("*"):
-                        archive.write(file_path, arcname = file_path.__str__().partition(self.path.name + os.path.sep)[2])
-                        self.logger.debug(f"Zipped '{file_path.name}'.")
-                
-                else: # If debug logging level is not set we can zip much FASTER!
-                    for file_path in self.path.rglob("*"):
-                        archive.write(file_path, arcname = file_path.__str__().partition(self.path.name + os.path.sep)[2])
+            with ZipFile(path_to_zip.absolute(), mode = "w") as archive:
+                for file_path in self.path.rglob("*"):
+                    archive.write(file_path, arcname = file_path.__str__().partition(self.path.name + os.path.sep)[2])
 
             self.logger.debug(f"Done zipping to '{path_to_zip.absolute()}'!")
-            
+
             # Updating path object.
             self.path = path_to_zip
